scripts/requirements.txt
mysql-connector-python~=8.0.28
python-dotenv~=0.19.2
names~=0.3.0